[DEV-6458] file organization update:
- moving fns in TimeFilter to helpers dir
- using existing mock data var mockSubmissions
 & updating its contents a bit
